+++ libraries/BL0942_SPI/CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.2] - 2026-01-28
+- fix negative values getWatt(), getVWave(), getIWave().
+- add **float getMaxCurrent()** **float getMaxCurrentRMS()**
+- add **float getMaxVoltage()** **float getMaxVoltageRMS()**
+- update calibration examples to include **getMax...()**
+- add **uint32_t errorCount()** to count readRegister CRC errors. (debug).
+- update readme.md
+- minor edits
+
 ## [0.1.1] - 2026-01-06
 - fix setUserMode() to force bit 0 and 1 to b1
 - add bit masks to getters and setters to get only valid bits.